submissions-team/pavan-kumar-reddy-kathi/requirements.txt
# Contains all libraries that are needed for this project
# pandas 2.2.3

pandas~=2.2.3
scipy~=1.15.2
matplotlib~=3.10.1
seaborn~=0.13.2
scikit-learn~=1.6.1
pyyaml~=6.0.2
numpy~=2.1.3
-e ./dpputility
dpputility~=0.1.0
xgboost~=3.0.0
setuptools~=75.8.0
packaging~=24.2
pillow~=11.1.0
zlib~=1.2.13
pip~=25.0
wheel~=0.45.1
protobuf~=5.29.4
openssl~=3.0.16
pytz~=2025.1
python-dateutil~=2.9.0.post0
pygments~=2.19.1
rich~=13.9.4
idna~=3.10
charset-normalizer~=3.4.1
threadpoolctl~=3.5.0
markupsafe~=3.0.2
joblib~=1.4.2
certifi~=2025.1.31
markdown~=3.7
cycler~=0.12.1
urllib3~=2.3.0
six~=1.17.0
contourpy~=1.3.1
fonttools~=4.56.0
python~=3.12.9
pyparsing~=3.2.1
kiwisolver~=1.4.8
werkzeug~=3.1.3
requests~=2.32.3
tensorflow~=2.19.0
lightgbm~=4.6.0
scikeras~=0.13.0